instagram_pipeline/requirements.txt
instaloader
openai-whisper
google-genai
python-dotenv
tqdm
